add concurrency; reduce number of params passed around;
