+++ package.json
@@ -22,6 +22,7 @@
     "lodash.template": "^4.5.0",
     "markdown-it": "^13.0.1",
     "node-fetch": "^2.6.6",
+    "p-map": "^5.5.0",
     "semver": "^7.3.7",
     "string-format": "^2.0.0"
   },
